requirements.txt
snowflake-connector-python==2.8.0
schema==0.7.5
PyYAML==6.0